+++ src/OpenTelemetry.Instrumentation.Owin/Implementation/OwinInstrumentationMetrics.cs
@@ -3,16 +3,15 @@
 
 using System.Diagnostics.Metrics;
 using System.Reflection;
+using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Instrumentation.Owin.Implementation;
 
 internal static class OwinInstrumentationMetrics
 {
-    internal static readonly AssemblyName AssemblyName = typeof(OwinInstrumentationMetrics).Assembly.GetName();
-
-    public static string MeterName => AssemblyName.Name;
-
-    public static Meter Instance => new(MeterName, SignalVersionHelper.GetVersion<OwinInstrumentationOptions>());
-
-    public static Histogram<double> HttpServerDuration => Instance.CreateHistogram<double>("http.server.request.duration", "s", "Duration of HTTP server requests.");
+    internal static readonly Assembly Assembly = typeof(OwinInstrumentationMetrics).Assembly;
+    internal static readonly AssemblyName AssemblyName = Assembly.GetName();
+    internal static readonly string MeterName = AssemblyName.Name;
+    internal static readonly Meter Instance = new(MeterName, Assembly.GetPackageVersion());
+    internal static readonly Histogram<double> HttpServerDuration = Instance.CreateHistogram<double>("http.server.request.duration", "s", "Duration of HTTP server requests.");
 }